This works very well for collecting all the data that we want on the DOF sensor

+++ FlightComputer/main/Sensor.h
@@ -4,7 +4,7 @@
 
 typedef struct per10mData
 {
-    uint8_t tenMeterFloor;
+    uint32_t tenMeterFloor;
     float   data;
 } per10mData_t;
